+++ jni/freedv/fdmdv.c
@@ -104,7 +104,7 @@ static COMP cadd(COMP a, COMP b)
 
 static float cabsolute(COMP a)
 {
-    return sqrt(pow(a.real, 2.0) + pow(a.imag, 2.0));
+    return sqrtf(powf(a.real, 2.0) + powf(a.imag, 2.0));
 }
 
 /*---------------------------------------------------------------------------*\
@@ -159,8 +159,8 @@ struct FDMDV * CODEC2_WIN32SUPPORT fdmdv_create(void)
            This helped PAPR for a few dB.  We don't need to adjust rx
            phase as DQPSK takes care of that. */
 	
-	f->phase_tx[c].real = cos(2.0*PI*c/(NC+1));
- 	f->phase_tx[c].imag = sin(2.0*PI*c/(NC+1));
+	f->phase_tx[c].real = cosf(2.0*PI*c/(NC+1));
+ 	f->phase_tx[c].imag = sinf(2.0*PI*c/(NC+1));
 
 	f->phase_rx[c].real = 1.0;
  	f->phase_rx[c].imag = 0.0;
@@ -179,18 +179,18 @@ struct FDMDV * CODEC2_WIN32SUPPORT fdmdv_create(void)
 
     for(c=0; c<NC/2; c++) {
 	carrier_freq = (-NC/2 + c)*FSEP + FDMDV_FCENTRE;
-	f->freq[c].real = cos(2.0*PI*carrier_freq/FS);
- 	f->freq[c].imag = sin(2.0*PI*carrier_freq/FS);
+	f->freq[c].real = cosf(2.0*PI*carrier_freq/FS);
+ 	f->freq[c].imag = sinf(2.0*PI*carrier_freq/FS);
     }
 
     for(c=NC/2; c<NC; c++) {
 	carrier_freq = (-NC/2 + c + 1)*FSEP + FDMDV_FCENTRE;
-	f->freq[c].real = cos(2.0*PI*carrier_freq/FS);
- 	f->freq[c].imag = sin(2.0*PI*carrier_freq/FS);
+	f->freq[c].real = cosf(2.0*PI*carrier_freq/FS);
+ 	f->freq[c].imag = sinf(2.0*PI*carrier_freq/FS);
     }
 	
-    f->freq[NC].real = cos(2.0*PI*FDMDV_FCENTRE/FS);
-    f->freq[NC].imag = sin(2.0*PI*FDMDV_FCENTRE/FS);
+    f->freq[NC].real = cosf(2.0*PI*FDMDV_FCENTRE/FS);
+    f->freq[NC].imag = sinf(2.0*PI*FDMDV_FCENTRE/FS);
 
     /* Generate DBPSK pilot Look Up Table (LUT) */
 
@@ -342,7 +342,7 @@ void tx_filter(COMP tx_baseband[NC+1][M], COMP tx_symbols[], COMP tx_filter_memo
     float   acc;
     COMP    gain;
 
-    gain.real = sqrt(2.0)/2.0;
+    gain.real = sqrtf(2.0)/2.0;
     gain.imag = 0.0;
     
     /*
@@ -518,7 +518,7 @@ void generate_pilot_fdm(COMP *pilot_fdm, int *bit, float *symbol,
 
     /* filter DPSK symbol to create M baseband samples */
 
-    filter_mem[NFILTER-1] = (sqrt(2)/2) * *symbol;
+    filter_mem[NFILTER-1] = (sqrtf(2)/2) * *symbol;
     for(i=0; i<M; i++) {
 	tx_baseband[i] = 0.0; 
 	for(j=M-1,k=M-i-1; j<NFILTER; j+=M,k+=M)
@@ -537,8 +537,8 @@ void generate_pilot_fdm(COMP *pilot_fdm, int *bit, float *symbol,
 
     for(i=0; i<M; i++) {
 	*phase = cmult(*phase, *freq);
-	pilot_fdm[i].real = sqrt(2)*2*tx_baseband[i] * phase->real;
-	pilot_fdm[i].imag = sqrt(2)*2*tx_baseband[i] * phase->imag;
+	pilot_fdm[i].real = sqrtf(2)*2*tx_baseband[i] * phase->real;
+	pilot_fdm[i].imag = sqrtf(2)*2*tx_baseband[i] * phase->imag;
     }
 }
 
@@ -557,7 +557,7 @@ void generate_pilot_fdm(COMP *pilot_fdm, int *bit, float *symbol,
 void generate_pilot_lut(COMP pilot_lut[], COMP *pilot_freq)
 {
     int   pilot_rx_bit = 0;
-    float pilot_symbol = sqrt(2.0);
+    float pilot_symbol = sqrtf(2.0);
     COMP  pilot_phase  = {1.0, 0.0};
     float pilot_filter_mem[NFILTER];
     COMP  pilot[M];
@@ -721,8 +721,8 @@ void CODEC2_WIN32SUPPORT fdmdv_freq_shift(COMP rx_fdm_fcorr[], COMP rx_fdm[], fl
 {
     int   i;
 
-    foff_rect->real = cos(2.0*PI*foff/FS);
-    foff_rect->imag = sin(2.0*PI*foff/FS);
+    foff_rect->real = cosf(2.0*PI*foff/FS);
+    foff_rect->imag = sinf(2.0*PI*foff/FS);
     for(i=0; i<nin; i++) {
 	*foff_phase_rect = cmult(*foff_phase_rect, *foff_rect);
 	rx_fdm_fcorr[i] = cmult(rx_fdm[i], *foff_phase_rect);
@@ -891,8 +891,8 @@ float rx_est_timing(COMP rx_symbols[],
        out single DFT at frequency 2*pi/P */
 
     x.real = 0.0; x.imag = 0.0;
-    freq.real = cos(2*PI/P);
-    freq.imag = sin(2*PI/P);
+    freq.real = cosf(2*PI/P);
+    freq.imag = sinf(2*PI/P);
     phase.real = 1.0;
     phase.imag = 0.0;
 
@@ -904,7 +904,7 @@ float rx_est_timing(COMP rx_symbols[],
     /* Map phase to estimated optimum timing instant at rate M.  The
        M/4 part was adjusted by experiment, I know not why.... */
     
-    rx_timing = atan2(x.imag, x.real)*M/(2*PI) + M/4;
+    rx_timing = atan2f(x.imag, x.real)*M/(2*PI) + M/4;
     
     if (rx_timing > M)
 	rx_timing -= M;
@@ -958,8 +958,8 @@ float qpsk_to_bits(int rx_bits[], int *sync_bit, COMP phase_difference[], COMP p
     int   msb=0, lsb=0;
     float ferr;
 
-    pi_on_4.real = cos(PI/4.0);
-    pi_on_4.imag = sin(PI/4.0);
+    pi_on_4.real = cosf(PI/4.0);
+    pi_on_4.imag = sinf(PI/4.0);
 
     /* Extra 45 degree clockwise lets us use real and imag axis as
        decision boundaries */
@@ -1026,8 +1026,8 @@ void snr_update(float sig_est[], float noise_est[], COMP phase_difference[])
     COMP  pi_on_4;
     int   c;
 
-    pi_on_4.real = cos(PI/4.0);
-    pi_on_4.imag = sin(PI/4.0);
+    pi_on_4.real = cosf(PI/4.0);
+    pi_on_4.imag = sinf(PI/4.0);
 
     /* mag of each symbol is distance from origin, this gives us a
        vector of mags, one for each carrier. */
@@ -1047,8 +1047,8 @@ void snr_update(float sig_est[], float noise_est[], COMP phase_difference[])
        quadrant for convenience. */
     
     for(c=0; c<NC+1; c++) {
-	refl_symbols[c].real = fabs(phase_difference[c].real);
-	refl_symbols[c].imag = fabs(phase_difference[c].imag);    
+	refl_symbols[c].real = fabsf(phase_difference[c].real);
+	refl_symbols[c].imag = fabsf(phase_difference[c].imag);    
 	n[c] = cabsolute(cadd(fcmult(sig_est[c], pi_on_4), cneg(refl_symbols[c])));
     }
      
@@ -1273,8 +1273,8 @@ float calc_snr(float sig_est[], float noise_est[])
    
     S = 0.0;
     for(c=0; c<NC+1; c++)
-	S += pow(sig_est[c], 2.0);
-    SdB = 10.0*log10(S+1E-12);
+	S += powf(sig_est[c], 2.0);
+    SdB = 10.0*log10f(S+1E-12);
     
     /* Average noise mag across all carriers and square to get an
        average noise power.  This is an estimate of the noise power in
@@ -1285,13 +1285,13 @@ float calc_snr(float sig_est[], float noise_est[])
     for(c=0; c<NC+1; c++)
 	mean += noise_est[c];
     mean /= (NC+1);
-    N50 = pow(mean, 2.0);
-    N50dB = 10.0*log10(N50+1E-12);
+    N50 = powf(mean, 2.0);
+    N50dB = 10.0*log10f(N50+1E-12);
 
     /* Now multiply by (3000 Hz)/(50 Hz) to find the total noise power
        in 3000 Hz */
 
-    N3000dB = N50dB + 10.0*log10(3000.0/RS);
+    N3000dB = N50dB + 10.0*log10f(3000.0/RS);
 
     snr_dB = SdB - N3000dB;
 
@@ -1457,7 +1457,7 @@ void CODEC2_WIN32SUPPORT fdmdv_get_rx_spectrum(struct FDMDV *f, float mag_spec_d
     /* window and FFT */
 
     for(i=0; i<2*FDMDV_NSPEC; i++) {
-	fft_in[i].real = f->fft_buf[i] * (0.5 - 0.5*cos((float)i*2.0*PI/(2*FDMDV_NSPEC)));
+	fft_in[i].real = f->fft_buf[i] * (0.5 - 0.5*cosf((float)i*2.0*PI/(2*FDMDV_NSPEC)));
 	fft_in[i].imag = 0.0;
     }
 
@@ -1465,12 +1465,12 @@ void CODEC2_WIN32SUPPORT fdmdv_get_rx_spectrum(struct FDMDV *f, float mag_spec_d
 
     /* FFT scales up a signal of level 1 FDMDV_NSPEC */
 
-    full_scale_dB = 20*log10(FDMDV_NSPEC);
+    full_scale_dB = 20*log10f(FDMDV_NSPEC);
 
     /* scale and convert to dB */
 
     for(i=0; i<FDMDV_NSPEC; i++) {
-	mag_spec_dB[i]  = 10.0*log10(fft_out[i].real*fft_out[i].real + fft_out[i].imag*fft_out[i].imag + 1E-12);
+	mag_spec_dB[i]  = 10.0*log10f(fft_out[i].real*fft_out[i].real + fft_out[i].imag*fft_out[i].imag + 1E-12);
 	mag_spec_dB[i] -= full_scale_dB;
     }
 }
